Trashening: Windows Chrome Edition
Y'all know that Windows NT 5.1 is Windows 2000, right? That Windows 95/98 runs only on museum pieces, VMs or outdated UA services? I'm padding Windows with old Chrome versions for now, I'm going to go back to it soon with IE, Opera and the Mozilla. I stopped accepting browser strings from beyond like 10 years ago because it's ridiculous. There's plenty of randomness in this list now and it's only going to grow, but those choices will be plausible.